// src/app/precisa/liquidacion/liquidacion_form/model/dtoExpedienteDetalle.ts
export class dtoExpedienteDetalle{

    constructor(){
        
    }
    
    IdExpediente?:           number;
    NroPeticion:            string;
    CodigoComponente:       string;
    Periodo:                string;
    MontoTotal?:             number;
    Copago?:                 number;
    Cantidad?:               number;
    IdAdmision?:             number;
    Linea?:                  number;
    Id?:                     number;
    IdSecuencia:            number;
    MontoCopagoVariable?:    number;
    IdSede?:                 number;
    Nombre:                   string;
    ESTADOdesc:             string;
    Estado?:                 number;    
    UsuarioCreacion?:       string;
    UsuarioModificacion?:   string;
    FechaCreacion:          Date;
    FechaModificacion:      Date;
    num?:                    number;
}